+++ getTeXRoot.py
@@ -7,12 +7,12 @@ if sublime.version() < '3000':
 	# we are on ST2 and Python 2.X
 	_ST3 = False
 	from latextools_utils.is_tex_file import is_tex_file
-	from latextools_utils.sublime import get_project_file_name
+	from latextools_utils.sublime_utils import get_project_file_name
 	from latextools_utils import parse_tex_directives
 else:
 	_ST3 = True
 	from .latextools_utils.is_tex_file import is_tex_file
-	from .latextools_utils.sublime import get_project_file_name
+	from .latextools_utils.sublime_utils import get_project_file_name
 	from .latextools_utils import parse_tex_directives
